set defaults for env variables and removed .env file

+++ startup.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
-ln -s /srv/mcpeserver/worlds/${WORLD}.cfg /srv/mcpeserver/${WORLD}.cfg
+if [ -f /srv/mcpeserver/worlds/${WORLD}.cfg ] && ! [ -f /srv/mcpeserver/${WORLD}.cfg ];
+then
+ ln -s /srv/mcpeserver/worlds/${WORLD}.cfg /srv/mcpeserver/${WORLD}.cfg
+fi
+
 systemctl enable mcpeserver@${WORLD}
 exec /lib/systemd/systemd
